Fixed minor bugs for compilation
