+++ 3d/C++/src/Frame.cpp
@@ -1,1 +1,13 @@
 #include "Frame.hpp"
+
+size_t Frame::size() {
+    return 0;
+}
+
+std::vector<cv::Mat> Frame::generateOutputImages() {
+    return std::vector<cv::Mat>();
+}
+
+std::vector<cv::Mat> Frame::generateOutputSynthImages() {
+    return std::vector<cv::Mat>();
+}
